requirements.txt
python-chess
matplotlib
requests